+++ resources/js/components/RoomMembershipButton.vue
@@ -69,7 +69,7 @@ const props = defineProps({
     required: true,
   },
   accessCode: {
-    type: Number,
+    type: String,
     default: null,
   },
   disabled: {