+++ resources/js/components/RoomTabRecordings.vue
@@ -323,7 +323,7 @@ const props = defineProps({
     required: true,
   },
   accessCode: {
-    type: Number,
+    type: String,
     default: null,
   },
   token: {