+++ tests/Backend/Unit/Console/helper/GreenlightRoom.php
@@ -21,7 +21,7 @@ class GreenlightRoom
     /**
      * GreenlightRoom constructor.
      */
-    public function __construct(string $id, string $user_id, string $name, string $uid, ?int $access_code = null, array $room_settings = [], bool $deleted = false)
+    public function __construct(string $id, string $user_id, string $name, string $uid, ?string $access_code = null, array $room_settings = [], bool $deleted = false)
     {
         $this->id = $id;
         $this->user_id = $user_id;